Task 32/appeal creation process in tg

+++ src/db/migrations/20240222165710_create_appeal_location_table.ts
@@ -23,8 +23,8 @@ const RelationRule = {
 export async function up(knex: Knex): Promise<void> {
     await knex.schema.createTable(TableName.APPEAL_LOCATIONS, function(table) {
         table.increments(ColumnName.ID).primary();
-        table.integer(ColumnName.LONGITUDE)
-        table.integer(ColumnName.LATITUDE)
+        table.float(ColumnName.LONGITUDE)
+        table.float(ColumnName.LATITUDE)
         table.string(ColumnName.ADDRESS)
         table
             .dateTime(ColumnName.CREATED_AT)
